Move the gl renderer to common
