+++ sdl/Cargo.toml
@@ -20,7 +20,6 @@ sdl2 = "0.38"
 crossbeam-channel = "0.5"
 cfg-if = "1.0"
 wav = "1.0"
-gl = "0.14"
 
 [features]
 default = ["static-sdl"]
